pyproject.toml
[build-system]
requires = ["setuptools", "numpy", "Cython"]
build-backend = "setuptools.build_meta"

[project]
name = "infinigen"
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version"]

description = "Infinite Photorealistic Worlds using Procedural Generation"
keywords = [
    "computer vision", 
    "data generation", 
    "procedural"
]
classifiers = [
    "Framework :: Blender", 
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10"
]

requires-python = "==3.10.*"
dependencies = [
    "bpy",
    "einops",
    "flow_vis",
    "frozendict",
    "geomdl",
    "gin_config>=0.5.0",
    "imageio",
    "matplotlib",
    "numpy==1.21.5",
    "opencv-python<=4.8.0.74",
    "psutil",
    "scikit-image==0.19.3",
    "scikit-learn",
    "scipy",
    "submitit",
    "tqdm",
    "trimesh==3.22.1",
    "zarr",
    "pandas",
    "landlab==2.4.1", 
    "pycparser==2.21", 
    "pyrender",
    "json5"
]

[project.optional-dependencies]
dev = [
    "pytest", 
    "pytest-ordering", 
    "pytest-cov", 
    "black", 
    "ruff", 
    "tabulate" # for integration test results
]
wandb = [
    "wandb"
]

[tool.setuptools]
# include-package-data is terribly named. package-data is still included if false, 
# just not the package-data setuptools would otherwise autogenerate from MANIFEST.in or version control
include-package-data = false 

[tool.setuptools.packages.find]
include = ["infinigen*"]

# must be specified as module paths with ., using classic filepaths with / will silently not do anything
exclude = [
    "infinigen.datagen.customgt.dependencies*",
    "infinigen.datagen.customgt.build*",
]

[tool.setuptools.package-data]

"*" = ["*.gin", "*.txt", "*.json"]

# Must be specified as paths relative to infinigen/ 
"infinigen" = [
    "terrain/**/*.soil", # extra files for SoilMachine
    "terrain/lib/**/*.so", # created by terrain compilation
    "terrain/lib/**/*.o", # created by terrain compilation
    "datagen/customgt/build/customgt", # created during opengl compilation
    "assets/creatures/parts/nurbs_data/*.npy", # stores creature nurbs centroids data, ideally will be deprecated
]

[tool.setuptools.dynamic]
version = {attr = "infinigen.__version__"}

[tool.pytest.ini_options]
testpaths = "tests"
junit_family = "xunit2"

[tool.cibuildwheel]
test-extras = ["dev"]
test-command = "pytest infinigen/tests"

[tool.ruff]
select = [
    "E", 
    "F", 
    #"B", 
    #"C901",
]
target-version = "py310"
line-length = 120 # will be reduced to 88 to match black soon

exclude = [
    "*/dependencies/"
]
ignore = [
    "E402", # we have lots of sys.path and opencv statements that must preceed imports, ignore for now
    "E501", # ignore line length for now
    "E701", # disable multi line statements until we enable black
    "F401", # unused imports are sometimes necessary for gin to work
    "E712", # some people think `if x == True` is more explicit, not going to force it
] 

[tool.ruff.per-file-ignores]
"__init__.py" = []
"infinigen/assets/*" = ["F841"] #transpiler currently creates lots of unused variables


[tool.ruff.mccabe]
max-complexity = 10